test(rust): pin the express-id REFERENCE-reader refusal at the u32 boundary (#3421)
Control/boundary/RED tests for content_hash.rs (parse_first_ref +
item_signature/sig_entity), step_text.rs (refs_in_line) and prepass.rs
(find_ifcproject_id): an ordinary id is unaffected, an id at exactly
u32::MAX parses, and an id above u32::MAX is refused rather than
wrapping onto a real low-numbered entity.

Extracted prepass.rs's inline test module to a sibling prepass_tests.rs
(house `_tests.rs` convention) to stay under the module-size ratchet
after the production fix's net line growth.

Refs #3421
